perf: lazy-load video card feedback panel
