+++ apps/web/src/components/video-card-feedback-menu.tsx
@@ -1,11 +1,12 @@
-import { useRef, useState } from "react";
-import { useAuth } from "../hooks/use-auth";
-import { useBlocked } from "../hooks/use-blocked";
-import { sendRecommendationFeedback } from "../lib/recommendation-tracker";
-import { goto } from "../lib/route-redirect";
+import { lazy, Suspense, useRef, useState } from "react";
 import type { VideoStream } from "../types/stream";
-import { RecommendationFeedbackDropdown } from "./recommendation-feedback-dropdown";
 import { MoreIcon } from "./watch-icons";
+
+const VideoCardFeedbackPanel = lazy(() =>
+  import("./video-card-feedback-panel").then((module) => ({
+    default: module.VideoCardFeedbackPanel,
+  })),
+);
 
 type Props = {
   stream: VideoStream;
@@ -14,41 +15,6 @@ type Props = {
 export function VideoCardFeedbackMenu({ stream }: Props) {
   const menuRef = useRef<HTMLButtonElement | null>(null);
   const [menuOpen, setMenuOpen] = useState(false);
-  const { isAuthed } = useAuth();
-  const { channels, videos, addChannel, removeChannel, addVideo, removeVideo } = useBlocked();
-  const channelBlocked =
-    !!stream.channelUrl &&
-    (channels.data ?? []).some((blocked) => blocked.url === stream.channelUrl);
-  const videoBlocked = (videos.data ?? []).some((blocked) => blocked.url === stream.id);
-
-  function requireAuth(): boolean {
-    if (isAuthed) return false;
-    goto("/");
-    return true;
-  }
-
-  function toggleVideoBlock() {
-    if (requireAuth()) return;
-    if (videoBlocked) {
-      removeVideo.mutate(stream.id);
-      return;
-    }
-    addVideo.mutate({ url: stream.id, global: false });
-  }
-
-  function toggleChannelBlock() {
-    if (!stream.channelUrl || requireAuth()) return;
-    if (channelBlocked) {
-      removeChannel.mutate(stream.channelUrl);
-      return;
-    }
-    addChannel.mutate({
-      url: stream.channelUrl,
-      name: stream.channelName,
-      thumbnailUrl: stream.channelAvatar,
-      global: false,
-    });
-  }
 
   return (
     <>
@@ -62,30 +28,13 @@ export function VideoCardFeedbackMenu({ stream }: Props) {
         <MoreIcon />
       </button>
       {menuOpen && (
-        <RecommendationFeedbackDropdown
-          stream={stream}
-          anchorEl={menuRef.current}
-          onClose={() => setMenuOpen(false)}
-          onNotInterested={() =>
-            sendRecommendationFeedback("not_interested", {
-              id: stream.id,
-              channelUrl: stream.channelUrl,
-            })
-          }
-          onLessFromChannel={
-            stream.channelUrl
-              ? () =>
-                  sendRecommendationFeedback("less_from_channel", {
-                    id: stream.id,
-                    channelUrl: stream.channelUrl,
-                  })
-              : undefined
-          }
-          onToggleVideoBlock={toggleVideoBlock}
-          onToggleChannelBlock={stream.channelUrl ? toggleChannelBlock : undefined}
-          videoBlocked={videoBlocked}
-          channelBlocked={channelBlocked}
-        />
+        <Suspense fallback={null}>
+          <VideoCardFeedbackPanel
+            stream={stream}
+            anchorEl={menuRef.current}
+            onClose={() => setMenuOpen(false)}
+          />
+        </Suspense>
       )}
     </>
   );
